# contracts/openapi/observatorio-v1.yaml
openapi: 3.0.3
info:
  title: Observatório APS
  version: v1
  description: >-
    Hand-written, not generated (see pom.xml: springdoc 2.8.6 targets Spring Framework 6 /
    Jakarta EE 10, incompatible with this project's Boot 4.1.1 / Framework 7 / EE 11). Checked
    against the routes Spring actually registers by OpenApiContractTest, in both directions.
    Routes and status codes beyond those the Tech Spec itself specifies (the whole /api/v1/auth/*
    and /api/v1/ready surface, the /api/v1/users/* administration routes, and every HTTP status
    code that is not explicitly 202/404 in §1.10) are PROJECT DECISIONS, documented in
    docs/adr/0008-superficie-http-de-autenticacao.md — never spec requirements. Likewise the
    scope-discovery reads GET /api/v1/runs, GET /api/v1/results/periods and the municipalities
    field of GET /api/v1/auth/me (docs/adr/0015-escopo-do-cliente-web-em-tempo-de-execucao.md).
security:
  - sessionCookie: []
...
paths:
...
  /api/v1/sources:
    get:
      tags: [sources]
      summary: >-
        Every source the caller holds MANAGE_SOURCE for, ordered by municipality then id. Narrowed,
        never refused: an installation-scoped grant sees all sources, a municipal one only its own,
        anyone else an empty list. No municipality parameter, because GET /auth/me lists only
        READ_CLINICAL municipalities and a technical admin holds none. Read-only: no recent reauth.
      responses:
        "200":
          description: OK. secretRef is a reference, never the secret value.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SourceResponse" }
    post:
      tags: [sources]
      summary: Register or update a PEC source configuration (MANAGE_SOURCE + recent reauth).
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      requestBody:
        required: true
        content:
          application/json:
            schema: { $ref: "#/components/schemas/CreateSourceRequest" }
      responses:
        "201":
          description: Registered. secretRef is a reference, never the resolved secret value.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceResponse" }
        "400":
          description: Source id already belongs to a different municipality.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/ApiError" }
        "404":
          description: Caller lacks MANAGE_SOURCE for the requested municipality.
  /api/v1/sources/{id}/test:
    post:
      tags: [sources]
      summary: Connectivity/capability/budget diagnostic against a registered source.
      security:
        - sessionCookie: []
          csrfCookie: []
          csrfHeader: []
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: >-
            Diagnostic result. detail never contains the secret. Every outcome but SOURCE_BUSY is
            stored as the source's lastDiagnostic for its current configuration version.
          content:
            application/json:
              schema: { $ref: "#/components/schemas/SourceTestResponse" }
        "404":
          description: Unknown source, or caller lacks MANAGE_SOURCE for its municipality.
  /api/v1/sources/{id}/requirements:
    get:
      tags: [sources]
      summary: >-
        The requirements a registered source meets, from its registration, its last diagnostic and
        the packaged PEC compatibility matrix (MANAGE_SOURCE; read-only, no recent reauth). Each
        code claims only what it checks — READ_CONNECTION means the last diagnostic under the
        current configuration was CONNECTED, not that the account is SELECT-only;
        PEC_VERSION_IN_MATRIX means the PEC version, read model and installation role appear in a
        VALIDATED entry, not the exact capability/PostgreSQL entry acquisition still requires.
      parameters:
        - name: id
          in: path
          required: true
          schema: { type: string }
      responses:
        "200":
          description: OK, in a fixed order.
          content:
            application/json:
              schema:
                type: array
                items: { $ref: "#/components/schemas/SourceRequirementResponse" }
        "404":
          description: Unknown source, or caller lacks MANAGE_SOURCE for its municipality.
...
components:
  securitySchemes:
    sessionCookie:
      type: apiKey
      in: cookie
      name: OBS_SESSION
      description: >-
        Opaque session token, never Spring's HttpSession-backed JSESSIONID. State-changing
        requests also require the CSRF double-submit pair: an XSRF-TOKEN cookie (set by any GET,
        e.g. /ready) echoed back in the X-XSRF-TOKEN header.
    csrfCookie:
      type: apiKey
      in: cookie
      name: XSRF-TOKEN
      description: >-
        CSRF token cookie issued by a GET such as /api/v1/ready. State-changing requests must
        echo its value in X-XSRF-TOKEN.
    csrfHeader:
      type: apiKey
      in: header
      name: X-XSRF-TOKEN
      description: CSRF token header matching the XSRF-TOKEN cookie.
...
  schemas:
    ApiError:
      type: object
      required: [code, message]
      properties:
        code: { type: string }
        message: { type: string }
...
    CreateSourceRequest:
      type: object
      required: [id, sourceFamily, host, port, databaseName, dbUser, secretRef, municipalityIbge]
      properties:
        id: { type: string }
        sourceFamily:
          type: string
          enum: [PEC_POSTGRESQL, EXTERNAL_DATASET]
          example: PEC_POSTGRESQL
        pecInstallationRole:
          type: string
          nullable: true
          enum: [PRONTUARIO, CENTRALIZADOR, UNKNOWN]
        sourceLocationKind:
          type: string
          nullable: true
          enum: [PRIMARY, READ_REPLICA, RESTORED_COPY, INSTITUTIONAL_DW]
        host: { type: string }
        port: { type: integer }
        databaseName: { type: string }
        dbUser: { type: string }
        secretRef: { type: string }
        municipalityIbge: { type: string }
        pecVersion: { type: string }
        readModel:
          type: string
          nullable: true
          enum: [PEC_DW, PEC_OLTP]
    SourceResponse:
      allOf:
        - $ref: "#/components/schemas/CreateSourceRequest"
        - type: object
          properties:
            sourceConfigurationVersion: { type: integer }
            createdAt: { type: string, format: date-time }
            lastDiagnostic:
              allOf:
                - $ref: "#/components/schemas/LastDiagnostic"
              nullable: true
              description: >-
                Null until a diagnostic ran against the current configuration version —
                re-registering a source hides the previous one. Always null from POST /sources.
    LastDiagnostic:
      type: object
      properties:
        outcome:
          type: string
          enum: [DESTINATION_NOT_ALLOWED, SOURCE_AUTHENTICATION_FAILED, SOURCE_PERMISSION_DENIED,
                 CONNECTION_FAILED, CONNECTED]
        detail:
          type: string
          nullable: true
          description: Never contains the secret (§1.12.7 L550).
        testedAt: { type: string, format: date-time }
    SourceRequirementResponse:
      type: object
      properties:
        code:
          type: string
          enum: [READ_CONNECTION, PEC_POSTGRESQL_FAMILY, PEC_VERSION_IN_MATRIX, MUNICIPAL_SCOPE]
        ok: { type: boolean }
    SourceTestResponse:
      type: object
      properties:
        outcome: { type: string }
        detail:
          type: string
          description: Never contains the secret (§1.12.7 L550).
        maxRows: { type: integer, format: int64 }
        maxDurationMs: { type: integer, format: int64 }
        statementTimeoutMs: { type: integer, format: int64 }
...